feat: target node 8, better code generation
BREAKING CHANGE: will no longer work on node 6

This was already set in the `engines` field, and in
`.nvmrc`, and CI, so is probably not too controversial.

As we have performance sensitive code which is async,
it's probably (untested) better to let the compiler
take advantage of the platform we actually support.