feat: return a list of RPM packages for easier processing
Packages up until now were returned as a single string with each new line containing the package data.
Readers of this output had to do further processing to extract the relevant package data.
This single-string format was added to the plugin in order to match what the snyk-docker-plugin was expecting; it was used for backwards-compatibility.

Now in addition to this we return the list of packages in a new field in the parser response so it is easier to process them.

Additionally changed the following:
- exceptions thrown by parsing the RPM package entries are no longer ignored
- exceptions are now allowed to throw to the caller of the parser
- improved docs around some types