refactor: Change return types
BREAKING CHANGE: changes the returned type from the extracting function
`getPackages`.  So far a string has been returned, which delimited each
entry in the DB with a newline, and each part of the entry with a tab.
Instead, we now pass a proper type back to the caller, making it easier
to have (and parse) optional fields.